+++ include/ListaAdjAresta.h
@@ -14,7 +14,7 @@ public:
 
     NoAresta* getCabeca();                                          // Retorna a cabeca da lista de adjacencia de arestas
     int getNumVerticesVizinhos();                                   // Retorna a quantidade de vertices vizinhos (vertices que estao na lista de arestas)
-    void adicionar_aresta(int origem, int destino, int peso);       // Adiciona uma aresta à lista
+    void adicionar_aresta(int origem, int destino, float peso);       // Adiciona uma aresta à lista
     void remover_aresta(int origem, int destino);                   // Remove uma aresta da lista
     void remover_primeira_aresta();                                 // Remove a primeira aresta da lista
 };
